+++ src/host_net/mod.rs
@@ -1,4 +1,7 @@
 #[cfg(feature = "iroh")]
 pub mod iroh_host;
+pub mod local_host;
 #[cfg(feature = "tcp")]
 pub mod tcp_host;
+#[cfg(feature = "usb")]
+pub mod usb_host;
