+++ src/piggui.rs
@@ -32,6 +32,7 @@ use crate::views::message_box::MessageRowMessage::ShowStatusMessage;
 use crate::views::ssid_dialog::SsidDialogMessage;
 #[cfg(feature = "usb")]
 use crate::views::ssid_dialog::SsidDialogMessage::HideSsidDialog;
+use host_net::usb_host;
 #[cfg(feature = "iroh")]
 use iroh_net::NodeId;
 #[cfg(any(feature = "iroh", feature = "tcp"))]
@@ -46,10 +47,7 @@ mod hardware_subscription;
 mod host_net;
 mod hw;
 mod hw_definition;
-pub mod local_device;
 mod net;
-#[cfg(feature = "usb")]
-mod usb;
 mod views;
 mod widgets;
 
@@ -124,7 +122,7 @@ fn main() -> iced::Result {
 #[allow(unused_variables)]
 fn reset_ssid(serial_number: String) -> Task<Message> {
     #[cfg(feature = "usb")]
-    return Task::perform(usb::reset_ssid_spec(serial_number), |res| match res {
+    return Task::perform(usb_host::reset_ssid_spec(serial_number), |res| match res {
         Ok(_) => InfoRow(ShowStatusMessage(Info(
             "Wi-Fi Setup reset to Default by USB".into(),
         ))),
