+++ src/views/ssid_dialog.rs
@@ -5,7 +5,7 @@ use self::SsidDialogMessage::{
 
 use crate::widgets::spinner::circular::Circular;
 use crate::widgets::spinner::easing::EMPHASIZED_ACCELERATE;
-use crate::{usb, Message};
+use crate::Message;
 use iced::keyboard::key;
 use iced::widget::{
     self, checkbox, column, container, horizontal_space, pick_list, row, text, text_input, Button,
@@ -15,6 +15,7 @@ use iced::{keyboard, Element, Event, Length, Task};
 use iced_futures::Subscription;
 use std::time::Duration;
 
+use crate::host_net::usb_host;
 use crate::hw_definition::description::{
     HardwareDetails, SsidSpec, SSID_NAME_MAX_LENGTH, SSID_PASS_MAX_LENGTH, SSID_PASS_MIN_LENGTH,
 };
@@ -57,7 +58,7 @@ pub enum SsidDialogMessage {
 #[allow(unused_variables)]
 fn send_ssid(serial_number: String, ssid_spec: SsidSpec) -> Task<Message> {
     #[cfg(feature = "usb")]
-    return Task::perform(usb::send_ssid_spec(serial_number, ssid_spec), |res| {
+    return Task::perform(usb_host::send_ssid_spec(serial_number, ssid_spec), |res| {
         Message::SsidSpecSent(res.map_err(|e| format!("Could not send SSID Spec: {}", e)))
     });
     #[cfg(not(feature = "usb"))]
